requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
environs==11.2.1
marshmallow==3.23.1
packaging==24.2
PyJWT==2.10.0
python-dotenv==1.0.1
sqlparse==0.5.2
